refactored separated out PiaFormat into PiaTypes and PiaUtils
